+++ ICD.Connect.Krang/Remote/Broadcast/IBroadcastHandler.cs
@@ -1,0 +1,19 @@
+using System;
+using ICD.Connect.Protocol.Network.Broadcast.Broadcasters;
+
+namespace ICD.Connect.Krang.Remote.Broadcast
+{
+	public interface IBroadcastHandler : IDisposable
+	{
+		/// <summary>
+		/// Gets the current broadcaster.
+		/// </summary>
+		IBroadcaster Broadcaster { get; }
+
+		/// <summary>
+		/// Sets the broadcaster for this handler.
+		/// </summary>
+		/// <param name="broadcaster"></param>
+		void SetBroadcaster(IBroadcaster broadcaster);
+	}
+}